ci: add desktop artifact builds
