+++ .github/workflows/desktop.yml
@@ -1,0 +1,101 @@
+name: Desktop Builds
+
+on:
+  push:
+    branches: [main, development]
+    tags:
+      - "v*"
+  pull_request:
+    branches: [main, development]
+  workflow_dispatch:
+
+permissions:
+  contents: read
+
+jobs:
+  desktop-build:
+    name: Desktop ${{ matrix.name }}
+    runs-on: ${{ matrix.platform }}
+    strategy:
+      fail-fast: false
+      matrix:
+        include:
+          - name: linux
+            platform: ubuntu-22.04
+            artifact: argentum-desktop-linux
+            args: ""
+            rust-targets: ""
+          - name: windows
+            platform: windows-latest
+            artifact: argentum-desktop-windows
+            args: ""
+            rust-targets: ""
+          - name: macos-aarch64
+            platform: macos-latest
+            artifact: argentum-desktop-macos-aarch64
+            args: "--target aarch64-apple-darwin"
+            rust-targets: "aarch64-apple-darwin"
+          - name: macos-x86_64
+            platform: macos-15-intel
+            artifact: argentum-desktop-macos-x86_64
+            args: "--target x86_64-apple-darwin"
+            rust-targets: "x86_64-apple-darwin"
+
+    steps:
+      - uses: actions/checkout@v4
+
+      - uses: actions/setup-node@v4
+        with:
+          node-version: "20"
+          cache: npm
+
+      - name: Install Rust stable
+        uses: dtolnay/rust-toolchain@stable
+        with:
+          targets: ${{ matrix.rust-targets }}
+
+      - name: Install Linux desktop dependencies
+        if: matrix.platform == 'ubuntu-22.04'
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y build-essential curl wget file libwebkit2gtk-4.1-dev libxdo-dev libssl-dev libayatana-appindicator3-dev librsvg2-dev patchelf
+
+      - name: Install dependencies
+        run: npm ci --legacy-peer-deps
+
+      - name: Verify version sync
+        run: npm run version:check
+
+      - name: Build TypeScript
+        run: npm run build
+
+      - name: Build desktop bundle
+        run: npm run desktop:build -- ${{ matrix.args }}
+
+      - name: Upload desktop artifacts
+        uses: actions/upload-artifact@v4
+        with:
+          name: ${{ matrix.artifact }}
+          if-no-files-found: error
+          path: |
+            src/desktop/target/release/bundle/**/*
+            src/desktop/target/*/release/bundle/**/*
+
+  desktop-release:
+    name: Attach desktop installers to release
+    runs-on: ubuntu-latest
+    needs: desktop-build
+    if: startsWith(github.ref, 'refs/tags/')
+    permissions:
+      contents: write
+    steps:
+      - uses: actions/download-artifact@v4
+        with:
+          path: artifacts/desktop
+
+      - uses: softprops/action-gh-release@v2
+        with:
+          overwrite_files: true
+          files: artifacts/desktop/**/*
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
